SYM-17: Separated unit and integration tests. Todo: Make them all pass, general clean up.
