+++ test/test_integration/res/expected_ansi_keylayout.keylayout
