+++ test/test_integration/res/expected_iso_keylayout.keylayout
