+++ test/test_integration/res/minimal_input_ansi.yaml
