+++ test/utils.py
@@ -1,0 +1,8 @@
+'''
+@author: Andrew J. Young
+@description: Objects which are useful for running tests.
+'''
+
+
+FILE_WRITERS_PATH = 'symboard.file_writers'
+
